Safari-611.1.10/LayoutTests/crypto/subtle/rsassa-pkcs1-v1_5-generate-key-sign-verify-expected.txt

Test sign/verify with RSASSA-PKCS1-v1_5 SHA-1 using a generated key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Generating key...
Signing...
PASS signature.byteLength is 256
Verifying...
PASS verified is true
PASS successfullyParsed is true

TEST COMPLETE
